pyproject.toml
[project]
name = "drone-factory"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "jinja2>=3.1.0",
  "jsonschema>=4.23.0",
  "pydantic>=2.11.0",
  "pytest>=8.4.0",
  "pyyaml>=6.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
